pyproject.toml
[project]
name = "tg-collector"
version = "0.1.0"
description = "ArchiveDrop: Telegram bot for collecting text and audio messages"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=1.1.1",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.1",
]
